refactor: PgQueue no longer takes a RetryPolicy
