+++ hook-consumer/src/main.rs
@@ -1,6 +1,6 @@
 use envconfig::Envconfig;
 
-use hook_common::pgqueue::{PgQueue, RetryPolicy};
+use hook_common::{pgqueue::PgQueue, retry::RetryPolicy};
 use hook_consumer::config::Config;
 use hook_consumer::consumer::WebhookConsumer;
 use hook_consumer::error::ConsumerError;
@@ -14,14 +14,9 @@ async fn main() -> Result<(), ConsumerError> {
         config.retry_policy.initial_interval.0,
         Some(config.retry_policy.maximum_interval.0),
     );
-    let queue = PgQueue::new(
-        &config.queue_name,
-        &config.table_name,
-        &config.database_url,
-        retry_policy,
-    )
-    .await
-    .expect("failed to initialize queue");
+    let queue = PgQueue::new(&config.queue_name, &config.table_name, &config.database_url)
+        .await
+        .expect("failed to initialize queue");
 
     let consumer = WebhookConsumer::new(
         &config.consumer_name,
